requirements.txt
pyyaml
jsonschema
requests
mkdocs
mkdocs-material
